带弹回效果的Indicator

+++ app/src/main/java/com/jackie/sample/utils/DensityUtils.java
@@ -6,23 +6,23 @@ import android.content.Context;
  * Created by Jackie on 2017/5/8.
  */
 public class DensityUtils {
-    public static int dp2px(Context context, float dpValue) {
+    public static int dp2px(Context context, double dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (dpValue * scale + 0.5f);
+        return (int) (dpValue * scale + 0.5);
     }
 
-    public static int px2dp(Context context, float pxValue) {
+    public static int px2dp(Context context, double pxValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (pxValue / scale + 0.5f);
+        return (int) (pxValue / scale + 0.5);
     }
 
-    public static int px2sp(Context context, float pxValue) {
+    public static int px2sp(Context context, double pxValue) {
         final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (pxValue / fontScale + 0.5f);
+        return (int) (pxValue / fontScale + 0.5);
     }
 
-    public static int sp2px(Context context, float spValue) {
+    public static int sp2px(Context context, double spValue) {
         final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (spValue * fontScale + 0.5f);
+        return (int) (spValue * fontScale + 0.5);
     }
 }
